refactor(ui): rewrite droid page to design system (Phase 4)
- Extract droid.tsx (749 LOC monolith) into pages/droid/ folder structure
- index.tsx: PageShell + PageHeader + ConfigLayout (left=SectionRail, form=DroidForm, json=RawJsonSettingsEditorPanel)
- droid-form.tsx: FormPane composing 3 FormSections (overview, byok, docs)
- sections/overview-section.tsx: Runtime install + Config files + Warnings
- sections/byok-section.tsx: Quick controls + Reasoning controls + BYOK summary + Custom models table
- sections/docs-section.tsx: Notes + Factory docs links + Provider docs links
- All state, handlers, and flows preserved verbatim — zero scope creep
- SectionRail anchors: overview / byok / docs
- Update dashboard-page-height-contract test to reference new folder path
- 14 unit tests added covering all sections, save flow, disabled states
